requirements.txt
# PyTorch with CUDA support
# For CUDA 12.1 (recommended for RTX 30xx/40xx)
# Install with: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
# For CUDA 11.8: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# For CPU only: pip install torch torchvision torchaudio
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Other dependencies
transformers>=4.37.0
accelerate>=0.26.0
qwen-omni-utils>=0.0.1
werpy>=0.0.1
flash-attn>=2.0.0
tqdm>=4.66.0
pyyaml>=6.0.0
soundfile>=0.12.0
librosa>=0.10.0
pandas>=2.0.0
numpy>=1.24.0

